third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-3/style-autocomplete-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that autocompletions are computed correctly when editing the Styles pane.


Running: testEmptyName

Running: testEmptyNameForce

Running: testSingleCharName

Running: testSubstringName

Running: testEmptyValue

Running: testImportantDeclarationDoNotToggleOnExclamationMark

Running: testImportantDeclaration

Running: testValueR

Running: testValueWithParenthesis

Running: testValuePrefixed

Running: testValueUnprefixed

Running: testValueSubstring

Running: testNameVariables

Running: testValueVariables
